fix: catch errors in user controller

+++ src/application/user/user.service.ts
@@ -7,6 +7,7 @@ import { IDatabaseConnection } from '../../persistence/database-connection.inter
 
 import { UserProperties, type User, type UserID } from './user'
 import { UserAlreadyExistError } from './user-already-exist.error'
+import { UserNotFoundError } from './user-not-found.error'
 import { IUserRepository } from './user.repository.interface'
 import { IUserService } from './user.service.interface'
 
@@ -63,12 +64,19 @@ export class UserService implements IUserService {
   }
   public async deleteById(id: UserID, currentUser: User): Promise<void> {
     return this.databaseConnection.transactional(async tx => {
+      const userExists = await this.repository.find(tx, id)
+
+      if (!userExists) throw new UserNotFoundError(id)
+
       if (currentUser.id === id)
         throw new ConflictException("You can't delete your own user account.")
-      await this.repository.deleteById(tx, id)
-      this.logger.verbose(
-        `User with id: ${id} has been deleted by admin: ${currentUser.id}`,
-      )
+
+      const userDeleted = await this.repository.deleteById(tx, id)
+
+      if (userDeleted)
+        this.logger.verbose(
+          `User with id: ${id} has been deleted by admin: ${currentUser.id}`,
+        )
     })
   }
 }
